server/requirements.txt
# FastAPI and server
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.5.2

# Database
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.13.2
greenlet>=3.0.0

# Twitter API
tweepy==4.16.0

# Trading API
alpaca-py==0.28.1

# Sentiment Analysis
vaderSentiment==3.3.2
transformers==4.45.0
torch==2.5.1
sentencepiece==0.2.0

# Utilities
python-dotenv==1.0.1
python-multipart==0.0.12
aiohttp==3.10.11
yfinance==0.2.40
pandas>=2.0.0

# Development
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
black==24.10.0
ruff==0.6.9

